# nestjs-backend/swagger-doc.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/portfolio/{id}": {
      "get": {
        "operationId": "PortfolioController_getByPortfolioId",
        "summary": "Get one Portfolio by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID of the portfolio to fetch",
            "example": "portfolio123",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The portfolio with the specified ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioDto"
                }
              }
            }
          }
        },
        "tags": [
          "Portfolio Service"
        ]
      }
    },
...
    "/core/portfolio/{portfolioId}": {
      "get": {
        "operationId": "CoreController_loadPortfolio",
        "summary": "Load Portfolio data of a specified portfolio",
        "parameters": [
          {
            "name": "portfolioId",
            "required": true,
            "in": "path",
            "description": "portfolio Id of portfolio",
            "example": "66d9ae695e15ad24b5e2053a",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Portfolio details by portfolio",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "portfolioId": "673edcfaa141dc2133e204e6",
                    "clientName": "IS484",
                    "portfolioName": "Final Presentation",
                    "portfolioAnalysis": {
                      "totalAssets": 119342.5,
                      "cashAmount": 24500.000000000022,
                      "securitiesValue": 94841.48474999999,
                      "dailyPL": -102.19,
                      "dailyPLPercentage": -0.11,
                      "totalPL": -657.5,
                      "totalPLPercentage": -0.69,
                      "annualizedRoR": 100
                    },
                    "triggeredAlerts": [
                      {
                        "id": "673dc25e23fe016af3c6fba5",
                        "ticker": "CAT",
                        "date": "2024-11-20T19:05:02.920Z",
                        "sentimentRating": 4,
                        "introduction": "Caterpillar...",
                        "assetName": "Caterpillar"
                      }
                    ],
                    "breachedRules": {
                      "breachedRules": [],
                      "recommendation": "",
                      "news": {
                        "buy": []
                      }
                    },
                    "portfolioBreakdown": {
                      "securities": [
                        {
                          "STOCK": 39.68
                        },
                        {
                          "BOND": 39.79
                        },
                        {
                          "CASH": 20.53
                        }
                      ],
                      "industry": [
                        {
                          "Construction and mining": 3.58
                        }
                      ],
                      "geography": [
                        {
                          "USA": 100
                        }
                      ]
                    },
                    "portfolioHoldings": [
                      {
                        "name": "Caterpillar",
                        "ticker": "CAT",
                        "type": "STOCK",
                        "geography": "USA",
                        "position": 9.021920934257132,
                        "market": 3394.046765598404,
                        "last": 376.20001220703125,
                        "cost": 387.510009765625,
                        "totalPL": -102.037903740274,
                        "totalPLPercentage": -2.92,
                        "dailyPL": -27.426441404574277,
                        "dailyPLPercentage": -0.8,
                        "positionsRatio": 3.58
                      }
                    ],
                    "orderExecutionProgress": []
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Core Service"
        ]
      }
    },
...
    "/generate/proposal": {
      "post": {
        "operationId": "PortfolioCreationController_generateOrders",
        "summary": "Generate orders for an optimized portfolio",
        "parameters": [
          {
            "name": "clientName",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "portfolioName",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "riskAppetite",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cash",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "managerId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "exclusions",
            "required": false,
            "in": "query",
            "description": "Array of the tickers of excluded assets",
            "example": [
              "AAPL"
            ],
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "minCash",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "maxCash",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Portfolio Generation Service"
        ]
      }
    },
    "/generate/{portfolioId}": {
      "get": {
        "operationId": "PortfolioCreationController_optimisePortfolio",
        "summary": "Optimise the current holdings of a portfolio",
        "parameters": [
          {
            "name": "portfolioId",
            "required": true,
            "in": "path",
            "description": "portfolio Id of portfolio",
            "example": "66d9ae695e15ad24b5e2053a",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Portfolio Generation Service"
        ]
      }
    },
...
  },
  "info": {
    "title": "Portfolio Optimizer Application API",
    "description": "core application in nestjs",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "AssetHoldingDto": {
        "type": "object",
        "properties": {
          "ticker": {
            "type": "string"
          },
          "cost": {
            "type": "number"
          },
          "quantity": {
            "type": "number"
          },
          "assetType": {
            "type": "string"
          }
        },
        "required": [
          "ticker",
          "cost",
          "quantity",
          "assetType"
        ]
      },
      "PortfolioDto": {
        "type": "object",
        "properties": {
          "client": {
            "type": "string"
          },
          "portfolioName": {
            "type": "string"
          },
          "riskAppetite": {
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH"
            ],
            "type": "string"
          },
          "cashAmount": {
            "type": "number"
          },
          "assetHoldings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AssetHoldingDto"
            }
          },
          "manager": {
            "type": "string"
          },
          "exclusions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "rules": {
            "type": "object"
          }
        },
        "required": [
          "client",
          "portfolioName",
          "riskAppetite",
          "cashAmount",
          "assetHoldings",
          "manager",
          "exclusions",
          "rules"
        ]
      },
...
    }
  }
}